requirements.txt
django-cors-headers==4.5.0

Django==4.0.3
psycopg2==2.9.3
django-environ==0.8.1
Pillow==9.1.0
django-phonenumber-field[phonenumbers] #https://github.com/stefanfoulis/django-phonenumber-field

#rest framwork
djangorestframework==3.13.1
djangorestframework-simplejwt==5.3.1
markdown==3.3.7
django-filter==22.1
gunicorn==20.1.0
django-imagekit==4.1.0

numpy==1.23.0
pandas==1.5.3

redis==5.0.7

nba_api==1.6.1
requests==2.32.3
beautifulsoup4==4.12.3

django-crontab==0.7.1
